feature-eng/requirements.txt
pandas>=2.0.0
pandas-ta>=0.3.14b
numpy>=1.24.0
influxdb-client>=1.36.0
python-dotenv>=1.0.0